scraper/requirements.txt
mongoengine==0.20.0
python-decouple==3.3